pyproject.toml
[tool.poetry]
name = "ssh-mock"
version = "0.3.1"
description = "Mocks an SSH Server"
authors = ["DanielHabenicht <daniel-habenicht@outlook.de>"]
include = ["mock-ssh/server_key"]
license = "MIT"
readme = "README.md"
exclude = [
    "tests",
]
repository = "https://github.com/DanielHabenicht/mock-ssh.git"
homepage = "https://github.com/DanielHabenicht/mock-ssh.git"
keywords = [
    "SSH",
    "Server",
    "Mocking",
    "Mock",
    "Testing",
]

[tool.poetry.scripts]
ssh-mock = "ssh_mock.cli:cli"

[tool.poetry.dependencies]
python = "^3.9"
paramiko = "3.5.1"
pyyaml = "^6.0.1"
jinja2 = "^3.1.2"

[tool.poetry.group.dev.dependencies]
python-semantic-release = "^8.3.0"
black = "^23.10.1"
flake8 = "^6.1.0"
mypy = "^1.6"
pylint = "^3.0.2"
isort = "^5.6.4"
pytest = "^8.2"
pytest-asyncio = "^0.26.0"
pytest-timeout = "^2.3.1"
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79

[tool.pylint."MESSAGES CONTROL"]
disable="""
    missing-module-docstring,
    missing-class-docstring,
    missing-function-docstring,
"""

[tool.semantic_release]
version_toml = [ "pyproject.toml:tool.poetry.version" ]
upload_to_repository = true
build_command = "poetry build"


[tool.semantic_release.publish]
dist_glob_patterns = ["dist/*"]
upload_to_vcs_release = true

[pytest]
timeout = 10

